datasets: side-effects moved to the client
